% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_utils.R
\name{add_steps}
\alias{add_steps}
\title{Add Steps to Recipe}
\usage{
add_steps(recipe, new_steps)
}
\arguments{
\item{recipe}{A \code{recipe}.}

\item{new_steps}{A series of step expressions. Use \code{expr} or \code{exprs} from \code{rlang}
to defuse steps.}
}
\description{
Add one or more steps to the end of a recipe object in sequence.
}
\examples{

\dontrun{
 .pred_date <- as.Date("2015-01-15")
 mtcars2 <- mtcars \%>\% mutate(hp_date = as.Date("2015-01-01"))
 mtcars2$hp_date[1:2] <- as.Date("2015-02-01")
 rec1 <- recipe(mtcars2, mpg ~ hp + hp_date)
 new_steps <- exprs(
   step_mutate(hp = ifelse(hp_date < .pred_date, hp, as.numeric(NA))),
   step_meanimpute(hp)
  )
 rec2 <- add_steps_to_recipe(rec1, new_steps)
}

}
