% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_utils.R
\name{extract_step_item}
\alias{extract_step_item}
\title{Extract step item}
\usage{
extract_step_item(recipe, step, item, enframe_item = TRUE)
}
\arguments{
\item{recipe}{Trained recipe object.}

\item{step}{Step from trained recipe.}

\item{item}{Item from trained recipe.}

\item{enframe_item}{Should the step item be enframed?}
}
\description{
Returns extracted step item from prepped recipe.
}
