% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample_utils.R
\name{fit_rsample_nested}
\alias{fit_rsample_nested}
\title{Fit models using nested split and recipe}
\usage{
fit_rsample_nested(
  split = NULL,
  recipe,
  model_func,
  strings_as_factors = FALSE,
  ...
)
}
\arguments{
\item{split}{An \code{rsplit} object created with \link{rolling_origin_nested}. If \code{recipe}
is trained with \code{prep(..., retain = TRUE)}, this argument is not needed and will not be
used.}

\item{recipe}{A trained or untrained recipe object. If not trained, \code{split} must be
included.}

\item{model_func}{A model function (ex: \code{lm} or \code{glm}). Must include arguments
\code{formula} and \code{data}.}

\item{strings_as_factors}{A logical: should character columns be converted to factors? This
affects the preprocessed training set (when retain = TRUE) as well as the results of
bake.recipe. Unlike \code{prep()}, the default is \code{FALSE}.}

\item{...}{Other arguments passed to \code{model_func}.}
}
\description{
This function makes it easy to fit a model using a nested split and a recipe object. A nested
split is one created with \code{rolling_origin_nested}, where each split is nested by a
time variable.
}
\details{
If \code{...} does not include \code{formula}, the formula will be extracted from \code{recipe}
using \code{formula(recipe)}.
}
