% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample_utils.R
\name{prepper_nested}
\alias{prepper_nested}
\title{Wrapper function for preparing recipes with nested resampling}
\usage{
prepper_nested(split, recipe, strings_as_factors = FALSE, ...)
}
\arguments{
\item{split}{An \code{rsplit} object created with \link{rolling_origin_nested}.}

\item{recipe}{An untrained \code{recipe} object.}

\item{strings_as_factors}{A logical: should character columns be converted to factors? This
affects the preprocessed training set (when retain = TRUE) as well as the results of
bake.recipe. Unlike \code{prep()}, the default is \code{FALSE}.}

\item{...}{Other arguments passed to \code{prep}.}
}
\description{
Makes it easier to prepare recipes using training data from nested resamples created with
\link{rolling_origin_nested}.
}
\details{
Sets the underlying \code{prep()} argument \code{fresh} to TRUE.
}
