% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_utils.R
\name{unnormalize}
\alias{unnormalize}
\title{Unnormalize variable}
\usage{
unnormalize(x, recipe, var)
}
\arguments{
\item{x}{Numeric vector to normalize.}

\item{recipe}{Trained recipe object.}

\item{var}{Variable name in the recipe object.}
}
\description{
Unormalizes variable using standard deviation and mean from a recipe object. See \code{?recipes}.
}
