% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample_utils.R
\name{rolling_origin_nested}
\alias{rolling_origin_nested}
\title{Rolling Origin Forecast Resampling with Irregular Timeslices}
\usage{
rolling_origin_nested(
  data,
  time_var,
  start = NULL,
  end = NULL,
  unit = "day",
  extend = FALSE,
  assess = 1,
  skip = 0,
  time_var_collapse = "round",
  ...
)
}
\arguments{
\item{data}{Data to use for training and prediction.}

\item{time_var}{A character. Name of date/time variable in \code{data}.}

\item{start}{A character. Minimum date/time for which predictions will be generated. If
\code{NULL}, then \code{min(time_var) + assess}}

\item{end}{A character. Maximum date/time for which predictions will be generated. If \code{NULL}
then \code{end = max(time_var)}.}

\item{unit}{A character string specifying a time unit or a multiple of a unit to nest the data.
Valid base units are \code{second, minute, hour, day, week, month, and year}. Arbitrary unique
English abbreviations as in the \code{period()} constructor are allowed. Rounding to multiple
of units (except weeks) is supported.}

\item{extend}{A logical. If \code{TRUE}, when \code{assess > 1} then all observations will be
predicted \code{assess} number of times. For example, if \code{data} has 10 observations after
nesting by \code{time_var}, \code{assess = 2}, all observations will be predicted 2 times when
\code{extend = TRUE}, whereas when \code{extend = FALSE} observations 10 and 3 (the default
\code{start}) will be predicted only once.}

\item{assess}{The number of nested date/times used for each assessment resample.}

\item{skip}{A integer indicating how many (if any) additional resamples to skip to thin the total
amount of data points in the analysis resample.}

\item{time_var_collapse}{One of \code{"floor"}, \code{"ceiling"}, or \code{"round"}. This
argument controls how \code{time_var} is collapsed prior to nesting. The options refer
to the respective \code{lubridate} functions. For example, see \code{?lubridate::floor_date}.}

\item{...}{Other arguments passed to \code{rsample::rolling_origin()}.}
}
\description{
Wrapper around \code{rsample::rolling_origin()} used to facilitate resampling of irregular time
slices.
}
\details{
Observations missing \code{time_var} are preserved and predicted last.
}
