% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_utils.R
\name{unnormalize}
\alias{unnormalize}
\title{Unnormalize variable}
\usage{
unnormalize(x, recipe, var, index = 1)
}
\arguments{
\item{x}{Numeric vector to normalize.}

\item{recipe}{Trained recipe object.}

\item{var}{Variable name in the recipe object.}

\item{index}{Determines which \code{step_normlize} to use.
If \code{step_normalize} is called twice, you will
need to specify \code{index = 2} to extract the information from the second
\code{step_normalize}.}
}
\description{
Unormalizes variable using standard deviation and mean from a recipe object. See \code{?recipes}.
}
